TDAR-6115 - tracking unique values... might need to dump it to a cache file to preserve memory

+++ processing/databaseprocessing/src/main/java/org/tdar/db/datatable/ImportColumn.java
@@ -28,7 +28,7 @@ public interface ImportColumn  {
 
     int compareToBySequenceNumber(ImportColumn b);
 
-    Set<Integer> getIntValues();
+    Set<Long> getIntValues();
 
     Set<Double> getFloatValues();
